+++ lib/savon.rb
@@ -5,6 +5,10 @@ require "savon/model"
 
 module Savon
 
+  def self.client(*args)
+    Client.new(*args)
+  end
+
   def self.configure
     yield config
   end