+++ spec/savon/savon_spec.rb
@@ -2,6 +2,10 @@ require "spec_helper"
 
 describe Savon do
 
+  it "provides a shortcut for creating a new client" do
+    Savon.client("http://example.com").should be_a(Savon::Client)
+  end
+
   it "memoizes the global config" do
     Savon.config.should equal(Savon.config)
   end